+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to MCP Gearbox will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [Unreleased]
+
+### Added
+- Added support for `mcp check` command to show list of installed agents on the user PC
+- Added support for `mcp rm` command to remove mcp servers
+- Added support for `mcp list` command to show configured mcp servers
+
+
 ## [0.0.9] - 2025-11-01
 
 ### Added